Trail FYP/requirements.txt
tensorflow>=2.8.0
torch>=1.10.0
scikit-learn>=1.0.0
pandas>=1.3.0
numpy>=1.21.0
fastapi>=0.68.0
uvicorn>=0.15.0
requests>=2.26.0
python-dotenv>=0.19.0
googlemaps>=4.6.0
